Move rank/file to index functions to utils module
