+++ utahchess/_board_repr.py
@@ -1,17 +1,7 @@
 from __future__ import annotations
 
 from utahchess.piece import Optional, Piece
-
-FILE_POSSIBILITIES = "abcdefgh"
-RANK_POSSIBILITIES = "87654321"
-
-
-def x_index_to_file(x: int) -> str:
-    return FILE_POSSIBILITIES[x]
-
-
-def y_index_to_rank(y: int) -> str:
-    return RANK_POSSIBILITIES[y]
+from utahchess.utils import x_index_to_file, y_index_to_rank
 
 
 def representation(_board: tuple[tuple[Optional[Piece], ...], ...]):
